+++ src/webview/src/components/tcsView.tsx
@@ -16,10 +16,11 @@
 // along with cph-ng.  If not, see <https://www.gnu.org/licenses/>.
 
 import Box from '@mui/material/Box';
-import React, { useState } from 'react';
+import { UUID } from 'crypto';
+import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { IProblem } from '@/types/types';
-import { msg } from '../utils';
+import { useProblemContext } from '../context/ProblemContext';
 import AcCongrats from './acCongrats';
 import CphFlex from './base/cphFlex';
 import ErrorBoundary from './base/errorBoundary';
@@ -32,9 +33,22 @@ interface TcsViewProps {
 
 const TcsView = ({ problem }: TcsViewProps) => {
   const { t } = useTranslation();
+  const { dispatch } = useProblemContext();
   const [draggedIdx, setDraggedIdx] = useState<number | null>(null);
   const [dragOverIdx, setDragOverIdx] = useState<number | null>(null);
   const [expandedStates, setExpandedStates] = useState<boolean[]>([]);
+  const [prevTcOrder, setPrevTcOrder] = useState<UUID[]>(problem.tcOrder);
+  const [focusTcId, setFocusTcId] = useState<UUID | null>(null);
+
+  useEffect(() => {
+    if (problem.tcOrder.length > prevTcOrder.length) {
+      const newIds = problem.tcOrder.filter((id) => !prevTcOrder.includes(id));
+      if (newIds.length === 1) {
+        setFocusTcId(newIds[0]);
+      }
+    }
+    setPrevTcOrder([...problem.tcOrder]);
+  }, [problem.tcOrder]);
 
   const handleDragStart = (idx: number, e: React.DragEvent) => {
     const states = problem.tcOrder.map((id) =>
@@ -71,7 +85,7 @@ const TcsView = ({ problem }: TcsViewProps) => {
     ) {
       const [movedId] = problem.tcOrder.splice(draggedIdx, 1);
       problem.tcOrder.splice(dragOverIdx, 0, movedId);
-      msg({ type: 'reorderTc', fromIdx: draggedIdx, toIdx: dragOverIdx });
+      dispatch({ type: 'reorderTc', fromIdx: draggedIdx, toIdx: dragOverIdx });
     }
 
     if (expandedStates.length > 0) {
@@ -140,13 +154,14 @@ const TcsView = ({ problem }: TcsViewProps) => {
                       onDragStart={(e) => handleDragStart(originalIdx, e)}
                       onDragEnd={handleDragEnd}
                       isDragging={draggedIdx === originalIdx}
+                      autoFocus={id === focusTcId}
                     />
                   </ErrorBoundary>
                 </Box>
               );
             })}
             <Box
-              onClick={() => msg({ type: 'addTc' })}
+              onClick={() => dispatch({ type: 'addTc' })}
               sx={{
                 minHeight: '40px',
                 cursor: 'pointer',