+++ src/webview/src/components/tips.tsx
@@ -23,7 +23,7 @@ import Alert from '@mui/material/Alert';
 import AlertTitle from '@mui/material/AlertTitle';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { msg } from '../utils';
+import { useProblemContext } from '../context/ProblemContext';
 import CphFlex from './base/cphFlex';
 import CphText from './base/cphText';
 import CphButton from './cphButton';
@@ -35,24 +35,25 @@ type TipMessage = {
 
 const Tips = () => {
   const { t } = useTranslation();
+  const { dispatch } = useProblemContext();
   const tipMessages: TipMessage[] = [
     { msg: t('tipMessage1') },
     { msg: t('tipMessage2') },
     { msg: t('tipMessage3') },
     {
       msg: t('tipMessage4'),
-      action: () => msg({ type: 'startChat' }),
+      action: () => dispatch({ type: 'startChat' }),
     },
     { msg: t('tipMessage5') },
     {
       msg: t('tipMessage6'),
       action: () =>
-        msg({ type: 'openSettings', item: 'cph-ng.sidebar.showAcGif' }),
+        dispatch({ type: 'openSettings', item: 'cph-ng.sidebar.showAcGif' }),
     },
     {
       msg: t('tipMessage7'),
       action: () =>
-        msg({
+        dispatch({
           type: 'openSettings',
           item: 'cph-ng.sidebar.hiddenStatuses',
         }),
@@ -61,7 +62,7 @@ const Tips = () => {
     {
       msg: t('tipMessage9'),
       action: () =>
-        msg({
+        dispatch({
           type: 'openSettings',
           item: 'cph-ng.compilation.useWrapper',
         }),